pyproject.toml
[project]
name = "jobsearch-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docx2pdf>=0.1.8",
    "docxtpl>=0.20.2",
    "fastapi>=0.128.0",
    "google-adk>=1.18.0",
    "ipykernel>=7.1.0",
    "ipython>=9.8.0",
    "langchain>=1.2.0",
    "langchain-azure-ai>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.5",
    "langchain-google-genai>=4.1.2",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "playwright>=1.57.0",
    "polars>=1.36.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
